backend-scraper/requirements.txt
Flask
Scrapy
scrapy-splash
pika
python-dotenv
six
pyjwt
celery
crochet


